+++ packages/engine/Source/Core/Credit.js
@@ -17,7 +17,7 @@ const creditToId = {};
  * @exception {DeveloperError} html is required.
  *
  * @example
- * //Create a credit with a tooltip, image and link
+ * // Create a credit with a tooltip, image and link
  * const credit = new Cesium.Credit('<a href="https://cesium.com/" target="_blank"><img src="/images/cesium_logo.png" title="Cesium"/></a>');
  */
 function Credit(html, showOnScreen) {